refactor: centralize ticket PDF export
